added login session and placeholder component for navbar
