+++ .env.sample
@@ -1,1 +1,2 @@
 DATABASE_URL=DATABASE_URL_WITH_USERNAME_AND_PASSWORD
+NEXTAUTH_SECRET=SECRET_FOR_JWT
