+++ components/navbar.tsx
@@ -1,0 +1,9 @@
+
+
+// TODO put changing sign in and sign out buttons here based on session
+// for sign out: `${window.location.origin}/`
+export default function Navbar() {
+  return (
+    <div>NAVBAR</div>
+  );
+}
