+++ types/next-auth.d.ts
@@ -1,0 +1,17 @@
+import NextAuth from 'next-auth';
+
+declare module 'next-auth' {
+  interface User {
+    username: string,
+  }
+
+  /**
+   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+   */
+  interface Session {
+    user: User
+    token: {
+      username: string,
+    }
+  }
+}
